валидация props
